+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/ViewModel/FullScreenPreviewVM.cs
@@ -1,4 +1,6 @@
-﻿using System.Windows.Input;
+﻿using System.Windows;
+using System.Windows.Input;
+using System.Windows.Media.Imaging;
 using ImoutoRebirth.Common.WPF;
 using ImoutoRebirth.Common.WPF.Commands;
 using ImoutoRebirth.Navigator.Services.Tags.Model;
@@ -17,6 +19,8 @@ internal class FullScreenPreviewVM : VMBase
     private IList<INavigatorListEntry>? _list;
     private IReadOnlyCollection<DelayItem>? _ugoiraFrameDelays;
     private IReadOnlyCollection<FileNote> _fileNotes = [];
+    private Size _viewPortSize;
+    private BitmapImage? _bitmapImage;
 
     public FullScreenPreviewVM(
         Func<INavigatorListEntry, Task<List<DelayItem>?>> ugoiraDelaysGetter, 
@@ -71,11 +75,27 @@ internal class FullScreenPreviewVM : VMBase
         get => _fileNotes;
         set => OnPropertyChanged(ref _fileNotes, value, () => FileNotes);
     }
-    
+
+    public Size ViewPortSize
+    {
+        get => _viewPortSize;
+        set
+        {
+            OnPropertyChanged(ref _viewPortSize, value, () => ViewPortSize);
+            OnPropertyChanged(nameof(Zoom));
+        }
+    }
+
+    public double Zoom => ViewPortSize.Width / _bitmapImage?.PixelWidth ?? 1;
+
     public async void SetCurrentEntry(INavigatorListEntry forEntry, IList<INavigatorListEntry> list)
     {
         Type = forEntry.Type;
         Path = forEntry.Path;
+
+        if (Type is ListEntryType.Image or ListEntryType.Png)
+            _bitmapImage = new BitmapImage(new Uri(Path));
+
         CurrentEntry = forEntry;
         _list = list;
 
@@ -84,7 +104,8 @@ internal class FullScreenPreviewVM : VMBase
         if (Type == ListEntryType.Ugoira)
             UgoiraFrameDelays = await _ugoiraDelaysGetter(forEntry);
 
-        FileNotes = await notesTask;
+        var notes = await notesTask;
+        FileNotes = notes.GroupBy(x => x.Source).MinBy(x => x.Key)?.ToList() ?? [];
         
         CurrentEntryNameChanged?.Invoke(this, System.IO.Path.GetFileName(forEntry.Path));
     }